+++ src/Schema/Definitions/DropDefinition.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Php\Support\Laravel\Database\Schema\Definitions;
+
+use Illuminate\Support\Fluent;
+
+/**
+ * @method Fluent cascade()
+ */
+class DropDefinition extends Fluent
+{
+}