+++ src/Schema/Definitions/LikeDefinition.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Php\Support\Laravel\Database\Schema\Definitions;
+
+use Illuminate\Support\Fluent;
+
+/**
+ * @method Fluent includingAll() INCLUDING ALL is an abbreviated form of INCLUDING DEFAULTS INCLUDING CONSTRAINTS
+ *     INCLUDING INDEXES INCLUDING STORAGE INCLUDING COMMENTS
+ */
+class LikeDefinition extends Fluent
+{
+}