+++ src/Schema/Postgres/Grammar.php
@@ -7,11 +7,13 @@ namespace Php\Support\Laravel\Database\Schema\Postgres;
 use Illuminate\Database\Schema\Grammars\PostgresGrammar;
 use Php\Support\Laravel\Database\Schema\Postgres\Grammar\CompressionModifier;
 use Php\Support\Laravel\Database\Schema\Postgres\Grammar\GrammarIndexes;
+use Php\Support\Laravel\Database\Schema\Postgres\Grammar\GrammarTable;
 use Php\Support\Laravel\Database\Schema\Postgres\Grammar\GrammarTypes;
 use Php\Support\Laravel\Database\Schema\Postgres\Grammar\GrammarViews;
 
 class Grammar extends PostgresGrammar
 {
+    use GrammarTable;
     use GrammarTypes;
     use GrammarIndexes;
     use GrammarViews;