+++ src/Schema/Postgres/Grammar/GrammarTable.php
@@ -1,0 +1,40 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Php\Support\Laravel\Database\Schema\Postgres\Grammar;
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Fluent;
+use Php\Support\Laravel\Database\Schema\Postgres\Compilers\CreateCompiler;
+
+trait GrammarTable
+{
+    public function compileCreate(Blueprint $blueprint, Fluent $command): string
+    {
+        $like        = $this->getCommandByName($blueprint, 'like');
+        $ifNotExists = $this->getCommandByName($blueprint, 'ifNotExists');
+
+        return CreateCompiler::compile(
+            $this,
+            $blueprint,
+            $this->getColumns($blueprint),
+            compact('like', 'ifNotExists')
+        );
+    }
+
+    /**
+     * Compile a drop table (if exists) command.
+     *
+     * @param Blueprint $blueprint
+     * @param Fluent $command
+     *
+     * @return string
+     */
+    public function compileDropIfExists(Blueprint $blueprint, Fluent $command)
+    {
+        $baseCompile = parent::compileDropIfExists($blueprint, $command);
+        $cascade     = $command->get('cascade') ? ' cascade' : '';
+        return "$baseCompile$cascade";
+    }
+}